define `plus` and `variable`
`plus` is the n-ary plus operation and `variable` is a symbol that
represents an unspecified quantity.

In order to avoid `plus` directly depending on `variable`, it uses the
concept `term` to restrict the template parameters of `plus`.

Each of the above is defined as it's own granular `cc_library` target in
`//sel`.  In addition, `//sel:contract` defines wrapper around
`assert` such that it is always checked.

Change-Id: Icd99ce63fb4f8dd46642ad3c6ebe0b774148614d