enable clangd and LSP settings for Zed (#2)
Add a target to generate a `compile_commands.json` file which can be
used with clangd. This file can be generated with

```
bazel run //tools:compile_commands
```

and will be generated in the root directory.

This file needs to be periodically regenerated, for example, when a new
translation unit is defined.

This commit also defines `.clangd` and `.zed` settings. `.clangd`
defines additional compilation options necessary for use of `clangd`
with the hermetic clang toolchain.

Change-Id: I02ae78c1110a7012611fcaa400adcfc592839c0f