speed up determination of lint targets in `check` workflow (#25)
Establish a new set of external dependency caches with the following
keys:
- Linux-default
- macOS-default
- Linux-local-config-cc

`Linux-local-config-cc` is created using `--config=local-config-cc` on
an Ubuntu runner where `local-config-cc` uses the local C++ toolchain
configuration.

Use of `--config=local-config-cc` allows the `clang-tidy-targets` job to
avoid the need to download the hermetic LLVM toolchain - any C++
toolchain is sufficient to perform `bazel cquery` in order to obtain a
list of targets.

Change-Id: I4dd3e1cdf3eed29442c62803710e70442dcabe14

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>